requirements.txt
flask
flask-cors
python-dateutil

requests
beautifulsoup4


openai
python-dotenv
